Timestep should have been an int
git-svn-id: file:///Users/petejw/Documents/libmesh_svn_bak@2908 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf

+++ trunk/libmesh/include/mesh/exodusII_io.h
@@ -108,7 +108,7 @@ class ExodusII_IO : public MeshInput<MeshBase>,
   
  private:
   ExodusII * ex_ptr;
-  double _timestep;
+  int _timestep;
 
 //-------------------------------------------------------------
   // local data
